Приложение authapp и основная маршрутизация
