+++ ittandem/authapp/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from .models import User
+
+admin.site.register(User)
