+++ ittandem/authapp/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class AuthappConfig(AppConfig):
+    name = 'authapp'
