+++ ittandem/ittandem/urls.py
@@ -14,10 +14,11 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.contrib import admin
-from django.urls import path
+from django.urls import path, include
 import mainapp.views as mainapp
 
 urlpatterns = [
     path('', mainapp.main),
     path('admin/', admin.site.urls),
+    path('user/', include('authapp.urls', namespace='user'))
 ]
