+++ ittandem/mainapp/models.py
@@ -1,23 +1,4 @@
 from django.db import models
-from django.contrib.auth.models import AbstractUser
-
-
-class User(AbstractUser):
-    avatar = models.ImageField(upload_to='user_avatars', blank=True, verbose_name='Аватар')
-    about_me = models.TextField(max_length=512, blank=True, verbose_name='О себе')
-    status = models.CharField(max_length=10, verbose_name='Статус пользователя', blank=True)
-
-
-class Contact(models.Model):
-    owner = models.ForeignKey(User, related_name='contacts', on_delete=models.CASCADE, verbose_name='Контакты')
-    link = models.TextField()
-
-
-class Project(models.Model):
-    owner = models.ForeignKey(User, related_name='projects', on_delete=models.CASCADE)
-    name = models.CharField(max_length=128, verbose_name='Название проекта')
-    desc = models.TextField(verbose_name='Описание проекта')
-    tags = models.ManyToManyField('Tag', related_name='tags')
 
 
 class Tag(models.Model):
@@ -30,6 +11,8 @@ class RelatedTag(models.Model):
     child_tag = models.ForeignKey(Tag, related_name='children', on_delete=models.CASCADE)
     parent = models.ForeignKey(Tag, related_name='parents', on_delete=models.CASCADE)
 
+
+from authapp.models import User
 
 class Skill(models.Model):
     user = models.ForeignKey(User, related_name='skills', on_delete=models.CASCADE)
