+++ ittandem/mainapp/templates/mainapp/base.html
@@ -41,7 +41,14 @@
             </div>
         </div>
     </header>
+
+    <div class="main">
+        {% block content %}
+
+        {% endblock %}
+    </div>
 </div>
+
 <footer id="sticky-footer" class="py-4 bg-dark text-white-50">
     <div class="container text-center">
         <small>Copyright &copy; Your Website</small>
